Introduce segment agent for single-data-file backup and restore
When backing up to or restoring from a remote filesystem, opening many file
handles can experience significant performance degradation.  To avoid this,
instead of re-opening the single data file for successive appends or reads
we create an intermediate pipe and tail the data into and out of the pipe,
so that we only incur the file-opening overhead once.

As a consequence, we now perform compression and decompression between the
intermediate pipe and the data file, rather than handling it in the COPY
TO PROGRAM call.

Signed-off-by: Jamie McAtamney <jmcatamney@pivotal.io>