Stop passing around objectCounts map
objectCounts keeps track of how many of each object we back up.
Previously we passed it into each function that set values in it, but at
some point it was made a global variable. We can clean up the functions
considerably by just using the global variable instead of passing it as
an argument. Usually we have concern about globals, but in this case it
seems fairly safe.

Author: Karen Huddleston <khuddleston@pivotal.io>